# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dobrenok Blog API",
        "description": "API для получения проектов и постов блога. Включает пагинацию, поиск, фильтрацию и сортировку.",
        "contact": {
            "name": "API Support",
            "email": "support@dobrenok.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        },
        {
            "url": "https://api.dobrenok.com/api/v1",
            "description": "Production server"
        }
    ],
    "paths": {
        "/projects/{id}/applications": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Создать заявку",
                "operationId": "2cbf2373ee3b4601d5bb84606c5be3c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID проекта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "data"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Обратный звонок"
                                    },
                                    "data": {
                                        "type": "object",
                                        "example": {
                                            "full_name": "Иван Иванов",
                                            "email": "ivan@example.com",
                                            "phone": "+79990000000"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Resource created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/projects/{id}/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Список сотрудников проекта",
                "description": "Возвращает пагинированный список сотрудников для указанного проекта",
                "operationId": "7ab30d47855cf2cc841abde4669e3d1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID проекта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1,
                            "example": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по имени или описанию",
                        "schema": {
                            "type": "string",
                            "example": "Иван"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Поле для сортировки",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "created_at",
                                "updated_at"
                            ],
                            "example": "created_at"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Направление сортировки",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Employee"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "pagination": {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Application": {
                "description": "Заявка",
                "properties": {
                    "id": {
                        "description": "ID заявки",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название заявки",
                        "type": "string",
                        "example": "Обратный звонок"
                    },
                    "project_id": {
                        "description": "ID проекта",
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "description": "Данные заявки (имя, email, телефон и т.д.)",
                        "type": "object"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Employee": {
                "description": "Сотрудник проекта",
                "properties": {
                    "id": {
                        "description": "ID сотрудника",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Имя сотрудника",
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "description": {
                        "description": "Описание сотрудника",
                        "type": "string",
                        "example": "Краткое био",
                        "nullable": true
                    },
                    "group": {
                        "description": "Группа сотрудника",
                        "type": "string",
                        "example": "Экспертная группа",
                        "nullable": true
                    },
                    "cover_photo_url": {
                        "description": "URL фотографии",
                        "type": "string",
                        "example": "http://localhost:8000/storage/employee-photos/photo.jpg"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 75
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/projects?page=2",
                        "nullable": true
                    },
                    "prev_page_url": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "not_found"
                            },
                            "message": {
                                "type": "string",
                                "example": "Resource not found"
                            },
                            "details": {
                                "type": "object",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}